refactor(gui): extract LogSyntaxHighlighter to its own header (#2554) — Principle VIII.
Lifted out of NetworkDiagnosticsDialog.cpp verbatim, by the same pure move as
TimeSeriesGraphWidget: the System Info dialog tails the same log file and wants
the same colouring, and a file-private class cannot be shared. No Q_OBJECT, so
no moc consequences. Nothing about the highlighter changed in the move.
— authored by agent (Claude Code) on behalf of @skerker
